backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.34.0
casparser==0.8.1
pyxirr==0.10.8
pandas==2.2.3
numpy==2.2.4
httpx==0.28.1
requests==2.32.5
diskcache==5.6.3
python-multipart==0.0.22
pydantic==2.12.5
pydantic-settings==2.8.1
anthropic==0.49.0
openai==1.69.0
google-generativeai==0.8.6
pdfminer.six==20240706
sse-starlette==2.2.1
mftool==3.2
